refactor: Split checker and generator directories
This patch adds individual subdirectories for the checker and generator
components within the compiler directory. The issue was that both parts
use similar concepts (e.g., scopes) but with differing implementations:
one focused on types, the other on values, which makes it impossible to
know whether a given source file is checker-related, generator-related,
or both.
