+++ packages/language/src/compiler/checker/checker.ts
@@ -1,29 +1,29 @@
 import type { SourceRange } from '@ast'
 import { ast } from '@ast'
 import type { Brand, Unit } from '@utility'
-import { getStandardModuleNames, getStandardModuleValue } from '../library/modules.js'
-import { CompoundError } from '../result/errors.js'
-import type { Result } from '../result/result.js'
-import { FunctionFacet } from '../type-system/base/function.js'
-import { ModuleFacet } from '../type-system/base/module.js'
-import { NumberFacet } from '../type-system/base/number.js'
-import { RecordFacet } from '../type-system/base/record.js'
-import { StringFacet } from '../type-system/base/string.js'
-import { BusFacet } from '../type-system/domain/bus.js'
-import { CurveFacet } from '../type-system/domain/curve.js'
-import { EffectFacet } from '../type-system/domain/effect.js'
-import { InstrumentFacet } from '../type-system/domain/instrument.js'
-import { ParameterFacet } from '../type-system/domain/parameter.js'
-import { PartFacet } from '../type-system/domain/part.js'
-import { PatternFacet } from '../type-system/domain/pattern.js'
-import { makeType, makeUnion } from '../type-system/factory.js'
-import type { Schema, SchemaItem } from '../type-system/schema.js'
-import type { FacetType, Type, Value } from '../type-system/types.js'
-import { BUS_NAMESPACE, busSchema, mixerSchema, partSchema, stepSchema, trackSchema } from './common.js'
-import { getCurveSegmentType } from './curves.js'
-import { CompileError } from './error.js'
+import { getStandardModuleNames, getStandardModuleValue } from '../../library/modules.js'
+import { CompoundError } from '../../result/errors.js'
+import type { Result } from '../../result/result.js'
+import { FunctionFacet } from '../../type-system/base/function.js'
+import { ModuleFacet } from '../../type-system/base/module.js'
+import { NumberFacet } from '../../type-system/base/number.js'
+import { RecordFacet } from '../../type-system/base/record.js'
+import { StringFacet } from '../../type-system/base/string.js'
+import { BusFacet } from '../../type-system/domain/bus.js'
+import { CurveFacet } from '../../type-system/domain/curve.js'
+import { EffectFacet } from '../../type-system/domain/effect.js'
+import { InstrumentFacet } from '../../type-system/domain/instrument.js'
+import { ParameterFacet } from '../../type-system/domain/parameter.js'
+import { PartFacet } from '../../type-system/domain/part.js'
+import { PatternFacet } from '../../type-system/domain/pattern.js'
+import { makeType, makeUnion } from '../../type-system/factory.js'
+import type { Schema, SchemaItem } from '../../type-system/schema.js'
+import type { FacetType, Type, Value } from '../../type-system/types.js'
+import { BUS_NAMESPACE, busSchema, mixerSchema, partSchema, stepSchema, trackSchema } from '../common.js'
+import { getCurveSegmentType } from '../curves.js'
+import { CompileError } from '../error.js'
 import { checkCyclicRoutings } from './routings.js'
-import { isSyntaxUnit, toBaseUnit } from './units.js'
+import { isSyntaxUnit, toBaseUnit } from '../units.js'
 
 export type CheckedProgram = Brand<ast.Program, 'language.CheckedProgram'>
 export type CheckResult = Result<CheckedProgram, CompoundError<CompileError>>
