+++ packages/language/src/compiler/generator/generator.ts
@@ -3,33 +3,33 @@ import type { Bus, BusId, InstrumentId, InstrumentRouting, Mixer, MixerRouting, 
 import { concatPatterns, createParallelPattern, createSerialPattern, mergePatterns, multiplyPattern } from '@core'
 import type { Numeric, Unit } from '@utility'
 import { numeric } from '@utility'
-import { getStandardModuleValue } from '../library/modules.js'
-import type { Function } from '../type-system/base/function.js'
-import { FunctionFacet } from '../type-system/base/function.js'
-import { ModuleFacet } from '../type-system/base/module.js'
-import { NumberFacet } from '../type-system/base/number.js'
-import { RecordFacet } from '../type-system/base/record.js'
-import { StringFacet } from '../type-system/base/string.js'
-import { BusFacet } from '../type-system/domain/bus.js'
-import { CurveFacet } from '../type-system/domain/curve.js'
-import { EffectFacet } from '../type-system/domain/effect.js'
-import { InstrumentFacet } from '../type-system/domain/instrument.js'
-import { ParameterFacet } from '../type-system/domain/parameter.js'
-import { PartFacet } from '../type-system/domain/part.js'
-import { PatternFacet } from '../type-system/domain/pattern.js'
-import { makeType } from '../type-system/factory.js'
-import { Numbers, Parameters } from '../type-system/helpers.js'
-import type { InferSchema, Schema } from '../type-system/schema.js'
-import type { FacetType, Value } from '../type-system/types.js'
-import type { CheckedProgram } from './checker.js'
-import { BUS_NAMESPACE, busSchema, partSchema, stepSchema, trackSchema } from './common.js'
-import type { CurveSegment as GeneratedCurveSegment } from './curves.js'
-import { createCurve, createCurveSegment, getCurveSegmentType, renderCurvePoints } from './curves.js'
-import { CompileError } from './error.js'
+import { getStandardModuleValue } from '../../library/modules.js'
+import type { Function } from '../../type-system/base/function.js'
+import { FunctionFacet } from '../../type-system/base/function.js'
+import { ModuleFacet } from '../../type-system/base/module.js'
+import { NumberFacet } from '../../type-system/base/number.js'
+import { RecordFacet } from '../../type-system/base/record.js'
+import { StringFacet } from '../../type-system/base/string.js'
+import { BusFacet } from '../../type-system/domain/bus.js'
+import { CurveFacet } from '../../type-system/domain/curve.js'
+import { EffectFacet } from '../../type-system/domain/effect.js'
+import { InstrumentFacet } from '../../type-system/domain/instrument.js'
+import { ParameterFacet } from '../../type-system/domain/parameter.js'
+import { PartFacet } from '../../type-system/domain/part.js'
+import { PatternFacet } from '../../type-system/domain/pattern.js'
+import { makeType } from '../../type-system/factory.js'
+import { Numbers, Parameters } from '../../type-system/helpers.js'
+import type { InferSchema, Schema } from '../../type-system/schema.js'
+import type { FacetType, Value } from '../../type-system/types.js'
+import type { CheckedProgram } from '../checker/checker.js'
+import { BUS_NAMESPACE, busSchema, partSchema, stepSchema, trackSchema } from '../common.js'
+import type { CurveSegment as GeneratedCurveSegment } from '../curves.js'
+import { createCurve, createCurveSegment, getCurveSegmentType, renderCurvePoints } from '../curves.js'
+import { CompileError } from '../error.js'
 import type { GenerateOptions } from './options.js'
 import type { GlobalScope, MutableNamespace, MutableScope, Scope } from './scopes.js'
 import { createGlobalScope, createLocalScope, createNamespace, resolveInScope } from './scopes.js'
-import { isSyntaxUnit, toNumberValue } from './units.js'
+import { isSyntaxUnit, toNumberValue } from '../units.js'
 
 /**
  * Generate a runnable program from an AST. This assumes the AST has already been
