+++ packages/language/src/index.ts
@@ -8,10 +8,10 @@ export * from './parser/parser.js'
 export * from './parser/error.js'
 export * from './parser/string.js'
 
-export type { GenerateOptions } from './compiler/options.js'
+export type { GenerateOptions } from './compiler/generator/options.js'
 export * from './compiler/error.js'
-export * from './compiler/checker.js'
-export * from './compiler/generator.js'
+export * from './compiler/checker/checker.js'
+export * from './compiler/generator/generator.js'
 
 export type { Documentation } from './library/documentation.js'
 export { getDocumentation } from './library/documentation.js'
