+++ packages/language/test/library/modules/effects.test.ts
@@ -1,8 +1,8 @@
 import { numeric } from '@utility'
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import type { GlobalScope } from '../../../src/compiler/scopes.js'
-import { createGlobalScope } from '../../../src/compiler/scopes.js'
+import type { GlobalScope } from '../../../src/compiler/generator/scopes.js'
+import { createGlobalScope } from '../../../src/compiler/generator/scopes.js'
 import { effectsModule } from '../../../src/library/modules/effects.js'
 import { RecordFacet } from '../../../src/type-system/base/record.js'
 import { EffectFacet } from '../../../src/type-system/domain/effect.js'
