+++ packages/language/test/library/modules/instruments.test.ts
@@ -2,12 +2,12 @@ import type { Envelope } from '@core'
 import { numeric } from '@utility'
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import type { GlobalScope } from '../../../src/compiler/scopes.js'
-import { createGlobalScope } from '../../../src/compiler/scopes.js'
-import { Numbers } from '../../../src/type-system/helpers.js'
+import type { GlobalScope } from '../../../src/compiler/generator/scopes.js'
+import { createGlobalScope } from '../../../src/compiler/generator/scopes.js'
 import { instrumentsModule } from '../../../src/library/modules/instruments.js'
 import { StringFacet } from '../../../src/type-system/base/string.js'
 import { InstrumentFacet } from '../../../src/type-system/domain/instrument.js'
+import { Numbers } from '../../../src/type-system/helpers.js'
 import { getFunctionExport } from './test-utils.js'
 
 function createFunctionContext (): GlobalScope {
