+++ packages/language/test/library/modules/patterns.test.ts
@@ -2,11 +2,11 @@ import { createSerialPattern, loopPattern, renderPatternEvents } from '@core'
 import { numeric } from '@utility'
 import assert from 'node:assert'
 import { describe, it } from 'node:test'
-import type { GlobalScope } from '../../../src/compiler/scopes.js'
-import { createGlobalScope } from '../../../src/compiler/scopes.js'
-import { Numbers } from '../../../src/type-system/helpers.js'
+import type { GlobalScope } from '../../../src/compiler/generator/scopes.js'
+import { createGlobalScope } from '../../../src/compiler/generator/scopes.js'
 import { patternsModule } from '../../../src/library/modules/patterns.js'
 import { PatternFacet } from '../../../src/type-system/domain/pattern.js'
+import { Numbers } from '../../../src/type-system/helpers.js'
 import { getFunctionExport } from './test-utils.js'
 
 function createFunctionContext (): GlobalScope {
